ChiralBoltz-main/ChiralBoltz-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chiralboltz"
version = "0.1.0"
description = "Mirror-image augmentation fine-tuning pipeline for Boltz-2 targeting heterochiral complex structure prediction"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "boltz>=0.4.0",
    "torch>=2.1.0",
    "pytorch-lightning>=2.2.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "rdkit>=2023.9.0",
    "pyyaml>=6.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4.0",
]
wandb = ["wandb>=0.17.0"]

[tool.setuptools.packages.find]
include = ["chiralboltz*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
